pyproject.toml
[tool.poetry]
name = "streamlit-agent"
version = "0.1.0"
description = ""
authors = ["Tim Conkling <tconkling@gmail.com>", "Joshua Carroll <joshua.carroll@snowflake.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "streamlit_agent"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = {extras = ["docarray"], version = "^0.0.234"}
openai = "^0.27.8"
streamlit = {path = "streamlit-1.24.1-py2.py3-none-any.whl"}
pypdf = "^3.12.2"
sentence-transformers = "^2.2.2"
torch = ">=2.0.0, !=2.0.1"
python-dotenv = "1.0.0"
psycopg2-binary = "2.9.6"
streamlit-modal = "0.1.0"
anthropic = "^0.3.0"
streamlit_extras = "0.2.7"
streamlit-supabase-auth = "1.0.2"
supabase="1.0.3"
storage3="0.5.3"
st-annotated-text="4.0.0"
pinecone-client="2.2.2"
tiktoken="0.4.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.4.1"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
